Grosse MAJ, désolé...

+++ src/models/models.py
@@ -12,6 +12,7 @@ from imblearn.metrics import classification_report_imbalanced
 
 from src.models.models_utils import call_memory, call_tqdm, call_dashboard, call_checkpoint, call_earlystopping, call_reducelr
 from src.generators.generator_fusion import FusionGenerator
+from src.generators.generator_image import ImageGenerator
 
 COLUMNS = [10,40,50,60,1140,1160,1180,1280,1281,1300,1301,1302,1320,1560,1920,1940,2060,2220,2280,2403,2462,2522,2582,2583,2585,2705,2905,]
 
@@ -27,13 +28,17 @@ class Model():
         save=True,
         report=True,
         summary=True,
+        predict=False,
         folder="src/models/",
+        name=None,
+        suffix=""
         ):
 
         self.preprocessor = None
         self.preprocessor_fitted = False
         self.model_fitted = False
         self.model = None
+        self.suffix=suffix
         
         self.generator = generator
         self.generator_test = generator_test
@@ -49,6 +54,9 @@ class Model():
         self.model_path = None
 
         return None
+
+    def get_name(self):
+        return self.name + self.suffix
 
     def fit_kwargs(self, validation_data=None):
 
@@ -80,7 +88,7 @@ class Model():
         return self.model_path if fold < 0 else self.model_fold_path
     
     def set_model_folder(self,):
-        self.model_path = Path(self.models_folder, self.type , self.name)
+        self.model_path = Path(self.models_folder, self.type , self.get_name())
         self.model_path.mkdir(parents=True, exist_ok=True)
         print(f"{self.model_path} created")
   
@@ -115,20 +123,28 @@ class Model():
             self.set_model_folder()
 
         if not fold < 0:
-            
             self.set_model_fold_folder(fold=fold)
-            path = Path(self.model_fold_path, 'model.joblib')
+            path = Path(self.model_fold_path)
         else :
-            path = Path(self.model_path, 'model.joblib')
+            path = Path(self.model_path)
 
-        print(f"model loaded from {path}")
-        return joblib.load(path)
+        if self.model_neural:
+            model = tf.keras.models.load_model(path)
+        else:
+            model = joblib.load(Path(path, "saved_model.pb"))
         
+        print(f"model loaded from {path}")
+        return model
 
     def save_model(self, model, in_fold=False):
         if self.save :
             path = self.model_path if not in_fold else self.model_fold_path
-            joblib.dump(model, Path(path, 'model.joblib'))
+            #joblib.dump(model, Path(path, 'model.joblib'))
+            
+            if self.model_neural:
+                model.save(Path(path))
+            else:
+                joblib.dump(model, Path(path, "saved_model.pb"))
             print(f"model saved in {path}")
 
     def fit_preprocessor(self, values):
@@ -150,13 +166,30 @@ class Model():
         else:
             return self.init_model()
 
-    def kfit(self, model=None, train_data=None, validation_data=None, fold=-1):
+
+    def unpack_iterator(self, iterator):
+        *unpacked, = iterator
+        x = []
+        y = []
+        for array in unpacked:
+            x.append(array[0])
+            y.append(array[1])
+        return np.concatenate(x), np.concatenate(y)
+
+    def kfit(self, model=None, train_data=None, validation_data=None, fold=-1, class_weight=None):
         
+        self.class_weight = class_weight if class_weight is not None else train_data.class_weight 
+
+        if self.type == "image" and isinstance(train_data, keras.preprocessing.image.DataFrameIterator):
+            fold = 0
+
         #If it's a fusion model, we build a special sequence with texts and images
         if self.type == "fusion" and not isinstance(train_data, FusionGenerator):
             train_data = FusionGenerator(train_data)
             validation_data = FusionGenerator(validation_data)
+
             print("data transformed to sequences for fusion")
+            fold = 0
 
         #If there is no model path, we create it
         if not self.model_path:
@@ -169,8 +202,6 @@ class Model():
         #Get the model
         model = self.get_model(fold=fold)
 
-        self.class_weight = train_data.class_weight
-
         #If fold is negtiv, we perform the cross validation process
         if fold < 0:
 
@@ -179,7 +210,7 @@ class Model():
 
             #Run the cross validation process
             for idx_fold, (train_index, valid_index) in enumerate(KFold(self.num_folds).split(train_data)):
-                
+
                 #Split
                 train_gen, valid_gen = train_data.split(split_indexes=[train_index, valid_index], is_batch=True)
 
@@ -189,10 +220,14 @@ class Model():
                     validation_data=valid_gen,
                     fold=idx_fold
                 )
-                self.save_model_summary(model)
-                
+                               
                 #Make prediction on the validation dataset
-                y_pred = self.predict(valid_gen, targets=valid_gen.targets, model=model, fold=idx_fold)
+                if self.model_neural:
+                    y_pred = self.predict(generator=valid_gen, targets=valid_gen.targets, model=model, fold=idx_fold)
+                else:
+                    valid_gen_unpacked, _ = self.unpack_iterator(valid_gen)
+                    y_pred = self.predict(features=valid_gen_unpacked, generator=valid_gen, targets=valid_gen.targets, model=model, fold=idx_fold)
+                    
                 y_true = valid_gen.decode(valid_gen.targets)
 
                 #Calculation of the recall score
@@ -202,38 +237,57 @@ class Model():
             #Select the best model an push it as attribute
             self.model = self.select_best_fold(scores)
             
-            #Save the best model as
+            #Save the best model in the main folder
             if not self.load:
                 self.save_model(self.model, in_fold=False)
 
             #Make prediction on the test dataset with the best model
-            y_pred = self.predict(validation_data, targets=validation_data.targets, model=self.model)
+            print("prediciton with best model")
+            if self.model_neural:
+                y_pred = self.predict(generator=validation_data, targets=validation_data.targets, model=self.model)
+            else:
+                validation_data_unpacked, _ = self.unpack_iterator(validation_data)
+                y_pred = self.predict(features=validation_data_unpacked, generator=validation_data, targets=validation_data.targets, model=model, fold=idx_fold)
+                    
+
+            
 
         #If fold is positiv, we make the fit
         else:
             print(f"starting fit process for fold {fold} : {self.type}")
 
             if not self.load:
-                #Fitting the model to the features
-                model.fit(
-                    train_data,
-                    **self.fit_kwargs(validation_data)
-                )
-                #model.summary(print_fn=self.save_model_summary)
+
+                #Fitting the model to the features for neural
+                if self.model_neural:
+                    model.fit(train_data, **self.fit_kwargs(validation_data))
+                #Fitting the model to the features for non neural
+                else:
+                    train_data, validation_data = self.unpack_iterator(train_data)
+                    model.fit(train_data, validation_data)
+                
                 print(f"end of fit in fold {fold}")
             
+                # Save model in the kfold folder
                 self.save_model(model, in_fold=True)
 
+                # Save the layers in the model summary file
+                self.save_model_summary(model)
+
+            self.model = model
             #Return the model
             return model
 
-    def predict(self, generator=None, targets=None, model=None, fold=-1):
+    def predict(self, features=None, generator=None, targets=None, model=None, fold=-1):
 
         #Get the model
         model = model if model is not None else self.model
 
         #Predict the targets
-        y_pred = model.predict(generator)
+        if features is not None:
+            y_pred = model.predict(features)
+        else:
+            y_pred = model.predict(generator)
         
         #Get the class if neural
         if self.model_neural:
@@ -251,13 +305,26 @@ class Model():
             #Get the classification report
             report = self.report_df(y_true, y_pred, fold=fold)
 
+            #Save the crosstab
+            self.crosstab(y_true, y_pred, fold=fold)
+
             #Save it in the summary if asked
             if self.summary:
                 self.save_summary(report)
         
         return y_pred
         
-        
+    def crosstab(self, y_true, y_pred, fold=-1):
+        #Calculate the crosstab
+        crosstab = pd.crosstab(
+            y_true, y_pred,
+            rownames=['Realité'], colnames=['Prédiction']
+        )
+        #Save it in a csv file
+        if self.save:
+            path = self.get_path(fold)
+            crosstab.to_csv(Path(path, f'crosstab_report.csv'), index= True)
+
     def report_df(self, y_true, y_pred, fold=-1):
         
         #Create clf report
@@ -290,13 +357,14 @@ class Model():
 
     def save_model_summary(self, model):
 
-        path = Path(self.model_path, "model_summary.txt")
-        if path.exists():
-            os.remove(path)
+        if self.model_neural:
+            path = Path(self.model_path, "model_summary.txt")
+            if path.exists():
+                os.remove(path)
 
-        with open(path, 'w') as f:
-
-            model.summary(print_fn=lambda x: f.write(x + '\n'))
+            #Save the model summary (layers) to a txt file
+            with open(path, 'w') as f:
+                model.summary(print_fn=lambda x: f.write(x + '\n'))
                 
 
     def save_summary(self, report, metrics=["precision", "recall", "f1-score"]):
@@ -324,7 +392,7 @@ class Model():
             df = pd.read_csv(summary_path, delimiter=",", index_col=0) if summary_path.exists() else pd.DataFrame(index=expected_columns)
 
             #Create or replace the scores obtained by the model
-            df[self.name] = expected_values
+            df[self.get_name()] = expected_values
             
             #Save it back to the summary file
             df.to_csv(summary_path)
